# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/v1/plans": {
      "get": {
        "operationId": "PlansController_listPlans",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "plan_ent",
                      "name": "Enterprise",
                      "slug": "enterprise",
                      "description": "Annual contract",
                      "type": "SUBSCRIPTION",
                      "channel": "SMS",
                      "billingInterval": "YEARLY",
                      "currency": "NGN",
                      "basePrice": "1200000",
                      "overagePrice": "0.9",
                      "overageEnabled": true,
                      "trialDays": 0,
                      "isActive": true,
                      "isPublic": true,
                      "monthlyQuota": 6000000,
                      "dailyQuota": 200000,
                      "isFree": false,
                      "maxContacts": 1000000,
                      "createdAt": "2026-03-08T10:47:11.873Z",
                      "updatedAt": "2026-09-24T10:47:11.873Z",
                      "_count": {
                        "subscriptions": 1
                      }
                    },
                    {
                      "id": "plan_growth",
                      "name": "Growth",
                      "slug": "growth",
                      "description": "For growing teams",
                      "type": "SUBSCRIPTION",
                      "channel": "EMAIL",
                      "billingInterval": "MONTHLY",
                      "currency": "NGN",
                      "basePrice": "50000",
                      "overagePrice": "1.8",
                      "overageEnabled": true,
                      "trialDays": 0,
                      "isActive": true,
                      "isPublic": true,
                      "monthlyQuota": 50000,
                      "dailyQuota": 5000,
                      "isFree": false,
                      "maxContacts": 25000,
                      "createdAt": "2025-08-20T10:47:11.873Z",
                      "updatedAt": "2026-09-24T10:47:11.873Z",
                      "_count": {
                        "subscriptions": 3
                      }
                    }
                  ],
                  "hasMore": false,
                  "nextCursor": null
                }
              }
            }
          }
        },
        "tags": [
          "Plans"
        ]
      },
      "post": {
        "operationId": "PlansController_upsertPlan",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Plans"
        ]
      }
    },
    "/admin/v1/plans/{id}": {
      "patch": {
        "operationId": "PlansController_updatePlan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertPlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Plans"
        ]
      }
    },
    "/admin/v1/plans/{id}/retire": {
      "post": {
        "operationId": "PlansController_retirePlan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Plans"
        ]
      }
    },
    "/admin/v1/plans/{id}/activate": {
      "post": {
        "operationId": "PlansController_activatePlan",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReasonDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Plans"
        ]
      }
    },
    "/admin/v1/products": {
      "get": {
        "operationId": "PlansController_listProducts",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "data": [
                    {
                      "id": "prod_email",
                      "name": "Email",
                      "channel": "EMAIL",
                      "description": "Transactional email",
                      "isActive": true,
                      "billingTypes": [
                        "CREDIT",
                        "SUBSCRIPTION"
                      ],
                      "createdAt": "2025-08-20T10:47:11.873Z",
                      "updatedAt": "2026-09-24T10:47:11.873Z",
                      "pricing": [
                        {
                          "id": "pp_email_ng",
                          "productId": "prod_email",
                          "countryCode": "NG",
                          "creditCost": "1.5",
                          "isActive": true,
                          "createdAt": "2025-08-20T10:47:11.873Z",
                          "updatedAt": "2026-09-24T10:47:11.873Z"
                        }
                      ]
                    },
                    {
                      "id": "prod_sms",
                      "name": "SMS",
                      "channel": "SMS",
                      "description": "Outbound SMS",
                      "isActive": true,
                      "billingTypes": [
                        "CREDIT",
                        "SUBSCRIPTION"
                      ],
                      "createdAt": "2025-08-20T10:47:11.873Z",
                      "updatedAt": "2026-09-24T10:47:11.873Z",
                      "pricing": [
                        {
                          "id": "pp_sms_gh",
                          "productId": "prod_sms",
                          "countryCode": "GH",
                          "creditCost": "9.5",
                          "isActive": true,
                          "createdAt": "2026-03-08T10:47:11.873Z",
                          "updatedAt": "2026-09-24T10:47:11.873Z"
                        },
                        {
                          "id": "11f63f12-861b-4653-b56b-dc68c63195b9",
                          "productId": "prod_sms",
                          "countryCode": "KE",
                          "creditCost": "7.25",
                          "isActive": true,
                          "createdAt": "2026-09-24T10:47:42.022Z",
                          "updatedAt": "2026-09-24T10:47:42.067Z"
                        }
                      ]
                    }
                  ],
                  "hasMore": false,
                  "nextCursor": null
                }
              }
            }
          }
        },
        "tags": [
          "Plans"
        ]
      },
      "post": {
        "operationId": "PlansController_createProduct",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Plans"
        ]
      }
    },
    "/admin/v1/products/{id}": {
      "patch": {
        "operationId": "PlansController_updateProduct",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "The channel is immutable; changing it means a different product.",
        "tags": [
          "Plans"
        ]
      }
    },
    "/admin/v1/pricing": {
      "get": {
        "operationId": "PlansController_listPricing",
        "parameters": [
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SMS",
                "EMAIL",
                "WHATSAPP",
                "VOICE",
                "VIDEO",
                "IN_APP",
                "FACE_LIVENESS",
                "MAPPING"
              ]
            }
          },
          {
            "name": "accountId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "asOf",
            "required": false,
            "in": "query",
            "description": "Defaults to today; pass a date to see the prices that applied then.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeHistory",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "asOf": "2026-09-24T10:54:34.899Z",
                  "data": [
                    {
                      "id": "pp_sms_gh",
                      "channel": "SMS",
                      "product": {
                        "id": "prod_sms",
                        "name": "SMS"
                      },
                      "countryCode": "GH",
                      "creditCost": "9.5",
                      "isActive": true,
                      "updatedAt": "2026-09-24T10:47:11.873Z",
                      "providerCostNgn": null,
                      "margin": null
                    },
                    {
                      "id": "11f63f12-861b-4653-b56b-dc68c63195b9",
                      "channel": "SMS",
                      "product": {
                        "id": "prod_sms",
                        "name": "SMS"
                      },
                      "countryCode": "KE",
                      "creditCost": "7.25",
                      "isActive": true,
                      "updatedAt": "2026-09-24T10:47:42.067Z",
                      "providerCostNgn": null,
                      "margin": null
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "Plans"
        ]
      },
      "post": {
        "operationId": "PlansController_upsertPricing",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProductPricingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Plans"
        ]
      }
    },
    "/admin/v1/provider-costs": {
      "get": {
        "operationId": "PlansController_listProviderCosts",
        "parameters": [
          {
            "name": "channel",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SMS",
                "EMAIL",
                "WHATSAPP",
                "VOICE",
                "VIDEO",
                "IN_APP",
                "FACE_LIVENESS",
                "MAPPING"
              ]
            }
          },
          {
            "name": "provider",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeHistory",
            "required": false,
            "in": "query",
            "description": "Without this only the currently effective rows are returned.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "example": {
                  "data": [],
                  "hasMore": false,
                  "nextCursor": null
                }
              }
            }
          }
        },
        "tags": [
          "Plans"
        ]
      },
      "post": {
        "operationId": "PlansController_upsertProviderCost",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProviderCostDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Plans"
        ]
      }
    },
    "/admin/v1/margin-targets": {
      "post": {
        "operationId": "PlansController_setMarginTarget",
        "parameters": [
          {
            "name": "x-csrf-token",
            "in": "header",
            "required": false,
            "description": "Required on every non-GET request. Read it from /auth/me.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Required on the write routes marked idempotent. One UUID per user intent, reused on every retry of that intent: the first call runs, later calls replay its response.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMarginTargetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Plans"
        ]
      }
    },
...
  },
  "info": {
    "title": "Zapiack Admin Console API",
    "description": "Backend for the Zapiack Admin Console.\n\n**Authentication is two steps.** `POST /auth/login` returns `totp_required` and\nsets a session cookie that every guarded route rejects until\n`POST /auth/2fa/verify` succeeds. Send `credentials: \"include\"` on every request.\n\n**Writes need `x-csrf-token`**, echoing the value from `/auth/2fa/verify` or\n`/auth/me`. Some also need an `Idempotency-Key` header and a written `reason`;\neach operation says which.\n\n**Every list returns `{ data, hasMore, nextCursor }`.** Extra context sits beside\n`data`, never instead of it.\n\n**Money is a decimal string** (`\"125000.0000\"`), never a number. Dates are\nAfrica/Lagos calendar days. Unknown query or body fields are rejected with 400.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "zpk_admin_at",
        "description": "Session cookie, set by the login flow. Not readable from JavaScript."
      }
    },
    "schemas": {
...
      "ReasonDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 8,
            "maxLength": 1000
          }
        },
        "required": [
          "reason"
        ]
      },
...
      "UpsertPlanDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 8,
            "maxLength": 1000
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "slug": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "description": {
            "type": "string",
            "maxLength": 400
          },
          "type": {
            "type": "string"
          },
          "channel": {
            "enum": [
              "EMAIL",
              "SMS"
            ],
            "type": "string"
          },
          "billingInterval": {
            "type": "string",
            "enum": [
              "MONTHLY",
              "YEARLY"
            ]
          },
          "basePrice": {
            "type": "string"
          },
          "overagePrice": {
            "type": "string"
          },
          "overageEnabled": {
            "type": "boolean"
          },
          "monthlyQuota": {
            "type": "string"
          },
          "dailyQuota": {
            "type": "string"
          },
          "maxContacts": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "isPublic": {
            "type": "boolean"
          },
          "isFree": {
            "type": "boolean"
          }
        },
        "required": [
          "reason",
          "name",
          "slug",
          "type",
          "channel",
          "basePrice",
          "overagePrice",
          "monthlyQuota"
        ]
      },
      "UpsertProductDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 8,
            "maxLength": 1000
          },
          "name": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "channel": {
            "enum": [
              "EMAIL",
              "SMS"
            ],
            "type": "string"
          },
          "description": {
            "type": "string",
            "maxLength": 400
          },
          "billingTypes": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "string",
              "enum": [
                "SUBSCRIPTION",
                "CREDIT"
              ]
            }
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "reason",
          "name",
          "channel",
          "billingTypes"
        ]
      },
      "UpsertProductPricingDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 8,
            "maxLength": 1000
          },
          "productId": {
            "type": "string"
          },
          "countryCode": {
            "type": "string",
            "minLength": 2,
            "maxLength": 2
          },
          "creditCost": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "required": [
          "reason",
          "productId",
          "countryCode",
          "creditCost"
        ]
      },
      "UpsertProviderCostDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 8,
            "maxLength": 1000
          },
          "provider": {
            "type": "string",
            "maxLength": 60
          },
          "channel": {
            "type": "string",
            "enum": [
              "SMS",
              "EMAIL",
              "WHATSAPP",
              "VOICE",
              "VIDEO",
              "IN_APP",
              "FACE_LIVENESS",
              "MAPPING"
            ]
          },
          "country": {
            "type": "string",
            "maxLength": 2
          },
          "network": {
            "type": "string",
            "maxLength": 60
          },
          "unitCostNgn": {
            "type": "string"
          },
          "effectiveFrom": {
            "type": "string"
          },
          "effectiveTo": {
            "type": "string"
          },
          "note": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "reason",
          "provider",
          "channel",
          "unitCostNgn",
          "effectiveFrom"
        ]
      },
      "SetMarginTargetDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "minLength": 8,
            "maxLength": 1000
          },
          "channel": {
            "type": "string",
            "enum": [
              "SMS",
              "EMAIL",
              "WHATSAPP",
              "VOICE",
              "VIDEO",
              "IN_APP",
              "FACE_LIVENESS",
              "MAPPING"
            ]
          },
          "targetMargin": {
            "type": "string",
            "description": "0.35 = 35%."
          },
          "effectiveFrom": {
            "type": "string"
          }
        },
        "required": [
          "reason",
          "channel",
          "targetMargin",
          "effectiveFrom"
        ]
      },
...
    }
  }
}